requirements.txt
# DKD Risk Prediction Model — Project Dependencies
# Install with: pip install -r requirements.txt

# Data processing & numerical computing
pandas>=2.0.0
numpy>=1.24.0

# Excel I/O
openpyxl>=3.0.0   # modern .xlsx format
xlrd>=2.0.1       # legacy .xls format support

# Machine learning & model training
xgboost>=3.0.0
scikit-learn>=1.3.0

# Visualization & explainability
matplotlib>=3.7.0
shap>=0.51.0

# API & web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6   # required for file uploads in FastAPI

# LLM & Semantic Analysis
google-generativeai>=0.8.0  # Gemini 1.5 Flash for column identification
python-dotenv>=1.0.0        # load .env files with API keys
